refactor(agent): extract formatToolProgress to tool_progress.go
Move the 146-line tool progress formatter from agent.go to its own
file for better organization. No logic changes.
