+++ agent/agent.go
@@ -2476,149 +2476,3 @@ func (a *Agent) ProcessDirect(ctx context.Context, content string) (string, erro
 // formatToolProgress generates a human-readable one-line summary of a tool call for progress display.
 // It parses the JSON args and extracts the most important parameter(s) based on the tool name.
 // Output is concise, max ~80 chars total.
-func formatToolProgress(name string, args string) string {
-	const maxLen = 80
-
-	// Helper to get a string field from parsed JSON
-	get := func(m map[string]any, key string) string {
-		if v, ok := m[key]; ok {
-			if s, ok := v.(string); ok {
-				return s
-			}
-			// Handle numeric types (e.g., limit as float64 from JSON)
-			return fmt.Sprintf("%v", v)
-		}
-		return ""
-	}
-
-	// Try to parse JSON args
-	var m map[string]any
-	parsed := json.Unmarshal([]byte(args), &m) == nil
-
-	// Helper to truncate and format the final result (rune-safe for multibyte chars)
-	truncate := func(s string, max int) string {
-		runes := []rune(s)
-		if len(runes) <= max {
-			return s
-		}
-		return string(runes[:max-3]) + "..."
-	}
-
-	if !parsed {
-		log.WithField("tool", name).WithField("raw_args", truncate(args, 200)).Debug("formatToolProgress: failed to parse tool args as JSON")
-	}
-
-	// Letta memory tools
-	switch name {
-	case "core_memory_append":
-		block := get(m, "block")
-		return truncate(fmt.Sprintf("core_memory_append: %s", block), maxLen)
-	case "core_memory_replace":
-		block := get(m, "block")
-		return truncate(fmt.Sprintf("core_memory_replace: %s", block), maxLen)
-	case "rethink":
-		block := get(m, "block")
-		return truncate(fmt.Sprintf("rethink: %s", block), maxLen)
-	case "archival_memory_insert":
-		return "archival_memory_insert"
-	case "archival_memory_search":
-		query := get(m, "query")
-		return truncate(fmt.Sprintf("archival_memory_search: %q", query), maxLen)
-	case "recall_memory_search":
-		query := get(m, "query")
-		startDate := get(m, "start_date")
-		endDate := get(m, "end_date")
-		parts := []string{}
-		if query != "" {
-			parts = append(parts, fmt.Sprintf("%q", query))
-		}
-		if startDate != "" || endDate != "" {
-			parts = append(parts, fmt.Sprintf("%s~%s", startDate, endDate))
-		}
-		return truncate(fmt.Sprintf("recall_memory_search: %s", strings.Join(parts, " ")), maxLen)
-	}
-
-	if !parsed {
-		// JSON parsing failed: show truncated raw args
-		raw := truncate(args, maxLen-len(name)-2)
-		if raw == "" {
-			return name
-		}
-		return truncate(fmt.Sprintf("%s: %s", name, raw), maxLen)
-	}
-
-	var summary string
-	switch name {
-	case "Shell":
-		summary = fmt.Sprintf("Shell: %s", get(m, "command"))
-	case "Read":
-		summary = fmt.Sprintf("Read: %s", get(m, "path"))
-	case "FileCreate":
-		summary = fmt.Sprintf("FileCreate: %s", get(m, "path"))
-	case "FileReplace":
-		summary = fmt.Sprintf("FileReplace: %s", get(m, "path"))
-	case "Grep":
-		pattern := get(m, "pattern")
-		path := get(m, "path")
-		include := get(m, "include")
-		target := path
-		if include != "" {
-			if target != "" {
-				target = include + " in " + target
-			} else {
-				target = include
-			}
-		}
-		if target != "" {
-			summary = fmt.Sprintf("Grep: %q in %s", pattern, target)
-		} else {
-			summary = fmt.Sprintf("Grep: %q", pattern)
-		}
-	case "Glob":
-		summary = fmt.Sprintf("Glob: %s", get(m, "pattern"))
-	case "WebSearch":
-		summary = fmt.Sprintf("WebSearch: %q", get(m, "query"))
-	case "Cron":
-		summary = fmt.Sprintf("Cron: %s", get(m, "action"))
-	case "SubAgent":
-		role := get(m, "role")
-		task := get(m, "task")
-		if role != "" {
-			summary = truncate(fmt.Sprintf("SubAgent [%s]: %s", role, task), maxLen)
-		} else {
-			summary = fmt.Sprintf("SubAgent: %s", task)
-		}
-	case "DownloadFile":
-		summary = fmt.Sprintf("DownloadFile: %s", get(m, "output_path"))
-	case "ChatHistory":
-		limit := get(m, "limit")
-		if limit != "" {
-			summary = fmt.Sprintf("ChatHistory: limit=%s", limit)
-		} else {
-			summary = "ChatHistory"
-		}
-	case "ManageTools":
-		action := get(m, "action")
-		mName := get(m, "name")
-		if mName != "" {
-			summary = fmt.Sprintf("ManageTools: %s %s", action, mName)
-		} else {
-			summary = fmt.Sprintf("ManageTools: %s", action)
-		}
-	case "card_create":
-		title := get(m, "title")
-		if title != "" {
-			summary = fmt.Sprintf("card_create: %q", title)
-		} else {
-			summary = "card_create"
-		}
-	default:
-		// Unknown tools (including MCP tools): show first 60 chars of args
-		raw := truncate(args, 60)
-		summary = fmt.Sprintf("%s: %s", name, raw)
-	}
-
-	// 去掉换行符，避免引用块断裂（工具参数可能含多行内容）
-	summary = strings.NewReplacer("\n", " ", "\r", "").Replace(summary)
-	return truncate(summary, maxLen)
-}
